==> deliverables/03_Split_Final.ipynb <==
# ADS-508-01-SP23 Team 8: Final Project

# Split and Preprocess ABT

Much of the code is modified from `Fregly, C., & Barth, A. (2021). Data science on AWS: Implementing end-to-end, continuous AI and machine learning pipelines. O’Reilly.`

## Install missing dependencies

[PyAthena](https://pypi.org/project/PyAthena/) is a Python DB API 2.0 (PEP 249) compliant client for Amazon Athena.

!pip install --disable-pip-version-check -q PyAthena==2.1.0
!pip install missingno

## Globally import libraries

import boto3
from botocore.client import ClientError
import sagemaker
import pandas as pd
import numpy as np
from pyathena import connect
from IPython.core.display import display, HTML
import missingno as msno
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import make_pipeline, Pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.feature_selection import VarianceThreshold
import datetime as dt
from io import BytesIO

%matplotlib inline

## Instantiate AWS SageMaker and S3 sessions

session = boto3.session.Session()
region = session.region_name
sagemaker_session = sagemaker.Session()
def_bucket = sagemaker_session.default_bucket()
bucket = 'sagemaker-us-east-ads508-sp23-t8'

s3 = boto3.Session().client(service_name="s3",
                            region_name=region)

print(f"Default bucket: {def_bucket}")
print(f"Public T8 bucket: {bucket}")

## Pass in ABT from CSV

s3_abt_csv_path = f"s3://{def_bucket}/team_8_data/abt/abt_encoded_df01.csv"
abt_encoded_df01 = pd.read_csv(s3_abt_csv_path)

### Perform train/test split

y01 = ['childpoverty']
abt_encoded_y01_vc01 = abt_encoded_df01[y01].to_numpy()
print(abt_encoded_y01_vc01.shape)
display(abt_encoded_y01_vc01[0:11])
abt_encoded_x01_df01 = abt_encoded_df01.drop(y01, axis=1)
print(abt_encoded_x01_df01.shape)
display(abt_encoded_x01_df01.head(11))

'''Citation for stratification by multiple cols:
https://stackoverflow.com/questions/45516424/sklearn-train-test-split-on-pandas-stratify-by-multiple-columns'''
abt_encoded_x01_df01['boroughs'] = abt_encoded_x01_df01['borough_bronx'].astype(int).astype(str) + abt_encoded_x01_df01['borough_brooklyn'].astype(int).astype(str) + abt_encoded_x01_df01['borough_manhattan'].astype(int).astype(str) + abt_encoded_x01_df01['borough_queens'].astype(int).astype(str) + abt_encoded_x01_df01['borough_staten island'].astype(int).astype(str)
display(abt_encoded_x01_df01.head(5))
train_x01, test_x01, train_y01, test_y01 = train_test_split(abt_encoded_x01_df01,
                                                            abt_encoded_y01_vc01,
                                                            test_size=.2,
                                                            stratify=abt_encoded_x01_df01[['boroughs']],
                                                            shuffle=True,
                                                            random_state=1699)

train_x01 = train_x01.drop(['boroughs', 'poverty'], axis=1)
test_x01 = test_x01.drop(['boroughs', 'poverty'], axis=1)

print(f'{train_x01.shape}')
print(f'{train_y01.shape}')
print(f'\n{test_x01.shape}')
print(f'{test_y01.shape}')

#### Examine featues with near zero variances

# Review near-zero variance (NZV) features for possible removal
train_x01_nzv_fit = VarianceThreshold(.025).fit(train_x01)
train_x01_nzv_vc01 = train_x01_nzv_fit.transform(train_x01)

# Get the names of the selected features
train_x01_nzv_fit_select_features = train_x01.columns[train_x01_nzv_fit.get_support()]

train_x01_nzv_df01 = pd.DataFrame(train_x01_nzv_vc01,
                                  columns=train_x01_nzv_fit_select_features)

display(train_x01_nzv_df01.head(5))
print(f'NZV transformed matrix dimensions = {train_x01_nzv_df01.shape}')

print(f'\n{train_x01.shape[1] - train_x01_nzv_df01.shape[1]} near zero variance features were eliminated')

print(train_x01.columns)
print(train_x01_nzv_df01.columns)

### Save ABT to S3

print(f'{train_x01.head(5)}')
print(f'{train_y01[0:5]}')
print(f'\n{test_x01.head(5)}')
print(f'{test_y01[0:5]}')

s3_train_x01_csv_path = f"s3://{def_bucket}/team_8_data/modeling_data/training/train_x01.csv"
train_x01.to_csv(s3_train_x01_csv_path,
                 index=False,
                 header=True)

s3_test_x01_csv_path = f"s3://{def_bucket}/team_8_data/modeling_data/testing/test_x01.csv"
test_x01.to_csv(s3_test_x01_csv_path,
                 index=False,
                 header=True)

# Define the S3 object key
train_y01_s3_key = 'team_8_data/modeling_data/training/train_y01.npy'

# Save the numpy array to S3
with BytesIO() as data:
    np.save(data, train_y01)
    data.seek(0)
    s3.upload_fileobj(data, def_bucket, train_y01_s3_key)

# Confirm that the numpy array was saved to S3
train_y01_response = s3.list_objects(Bucket=def_bucket,
                           Prefix=train_y01_s3_key)
print(train_y01_response)

# Define the S3 object key
test_y01_s3_key = 'team_8_data/modeling_data/testing/test_y01.npy'

# Save the numpy array to S3
with BytesIO() as data:
    np.save(data, test_y01)
    data.seek(0)
    s3.upload_fileobj(data, def_bucket, test_y01_s3_key)

# Confirm that the numpy array was saved to S3
test_y01_response = s3.list_objects(Bucket=def_bucket,
                           Prefix=test_y01_s3_key)
print(test_y01_response)

## Release Resources

%%html

<p><b>Shutting down your kernel for this notebook to release resources.</b></p>
<button class="sm-command-button" data-commandlinker-command="kernelmenu:shutdown" style="display:none;">Shutdown Kernel</button>
        
<script>
try {
    els = document.getElementsByClassName("sm-command-button");
    els[0].click();
}
catch(err) {
    // NoOp
}    
</script>

%%javascript

try {
    Jupyter.notebook.save_checkpoint();
    Jupyter.notebook.session.delete();
}
catch(err) {
    // NoOp
}
